requirements.txt
# Python Dependencies for Local Document Extraction Copilot
# Last updated: 2026-01-13
# Usage: pip install -r requirements.txt

# ============================================================
# CORE BACKEND
# ============================================================
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ============================================================
# DATABASE & ORM
# ============================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pgvector==0.2.0

# ============================================================
# LLM & EMBEDDINGS
# ============================================================
# ollama client is installed separately or via system package
sentence-transformers==2.2.2
numpy==1.26.3  # Updated for Python 3.12+ compatibility

# ============================================================
# OBSERVABILITY & TELEMETRY
# ============================================================
opentelemetry-api==1.20.0  # Python 3.12 compatible
opentelemetry-sdk==1.20.0  # Python 3.12 compatible
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0

# ============================================================
# UTILITIES
# ============================================================
python-dotenv==1.0.0
httpx==0.25.2
pydantic-core==2.14.1
pdfplumber==0.11.0
pytesseract==0.3.10

# ============================================================
# DEVELOPMENT & TESTING
# ============================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8
mypy==1.7.1
